# share/htdocs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  api.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
      username:
        type: string
    type: object
  api.ConfigSetRequest:
    properties:
      key:
        type: string
      namespace:
        type: string
      value: {}
    type: object
...
  api.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  api.FeatureFlagRequest:
    description: Feature flag update request
    properties:
      enabled:
        example: true
        type: boolean
      flag:
        example: new_ui
        type: string
    type: object
...
  api.ResetPasswordRequest:
    properties:
      password:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  api.StatusResponse:
    description: Status message
    properties:
      status:
        example: ok
        type: string
    type: object
...
  api.SuccessResponse:
    description: Generic success response
    properties:
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  models.Entity:
    properties:
      content:
        items:
          type: integer
        type: array
      created_at:
        description: Timestamps as nanosecond epoch for efficiency
        type: integer
      id:
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: integer
    type: object
host: localhost:8085
info:
  contact:
    email: support@entitydb.io
    name: EntityDB Support
  description: A temporal database with pure entity-based architecture
  license:
    name: MIT
    url: https://github.com/osakka/entitydb/blob/main/LICENSE
  termsOfService: https://github.com/osakka/entitydb
  title: EntityDB API
  version: 2.12.0
paths:
...
  /api/v1/config:
    get:
      consumes:
      - application/json
      description: Retrieve current system configuration
      parameters:
      - description: Configuration namespace
        in: query
        name: namespace
        type: string
      - description: Configuration key
        in: query
        name: key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Entity'
            type: array
      security:
      - BearerAuth: []
      summary: Get configuration
      tags:
      - configuration
  /api/v1/config/set:
    post:
      consumes:
      - application/json
      description: Update system configuration values
      parameters:
      - description: Configuration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ConfigSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Set configuration
      tags:
      - configuration
...
  /api/v1/feature-flags:
    get:
      consumes:
      - application/json
      description: Retrieve all feature flags
      parameters:
      - description: Stage filter (alpha, beta, stable)
        in: query
        name: stage
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Entity'
            type: array
      security:
      - BearerAuth: []
      summary: Get feature flags
      tags:
      - configuration
  /api/v1/feature-flags/set:
    post:
      consumes:
      - application/json
      description: Update a feature flag value
      parameters:
      - description: Feature flag data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.FeatureFlagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Entity'
      security:
      - BearerAuth: []
      summary: Set feature flag
      tags:
      - configuration
...
  /api/v1/users/change-password:
    post:
      consumes:
      - application/json
      description: Change the current user's password
      parameters:
      - description: Password change data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
      summary: Change user's own password
      tags:
      - users
  /api/v1/users/create:
    post:
      consumes:
      - application/json
      description: Create a new user entity with authentication credentials
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Entity'
      summary: Create a new user
      tags:
      - users
  /api/v1/users/reset-password:
    post:
      consumes:
      - application/json
      description: Reset a user's password (requires admin permission)
      parameters:
      - description: Password reset data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
      summary: Reset user password (admin only)
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Example: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"